update delete deployment

+++ src/utils/deleteDeployments.ts
@@ -15,42 +15,20 @@ export default async (
       environment: environment
     });
 
-    // const deployments = await octokit.graphql(`
-    //   query GetDeployments($owner: String!, $repo: String!, $environments: [String!]) {
-    //     repository(owner: $owner, name: $repo) {
-    //       deployments(first: 100, environments: $environments) {
-    //         nodes {
-    //           id
-    //         }
-    //       }
-    //     }
-    //   }`, { ...repo, environments: [environment] })
-
-    // const nodes = deployments.repository?.deployments?.nodes;
-
     console.log(JSON.stringify(deployments))
 
-    for (const deployment of deployments.data) {  
-      await octokit.repos.deleteDeployment({
-        ...repo,
-        deployment_id: deployment.id
-      })
+    const existing = deployments.data.length;
+    if (existing < 1) {
+        console.log(`No exiting deployments found for pull request`);
+        return;
+    }
+
+    for (const deployment of deployments.data) {
+        console.log(`Deleting existing deployment - ${deployment.id}`);
+
+        await octokit.repos.deleteDeployment({
+            ...repo,
+            deployment_id: deployment.id,
+        });
     };
-    
-    // if (nodes.length <= 0) {
-    //     console.log(`No exiting deployments found for pull request`);
-    //     return;
-    // }
-
-    // for (const node of nodes) {
-    //     console.log(`Deleting existing deployment - ${node.id}`);
-
-    //     await octokit.graphql(`
-    //       mutation DeleteDeployment($id: ID!) {
-    //         deleteDeployment(input: {id: $id} ) {
-    //           clientMutationId
-    //         }
-    //       }
-    //     `, { id: node.id })
-    // }
 }
